Reworked A/D conversions

+++ src/core/adc.h
@@ -2,7 +2,7 @@
 #define adc_h
 
 void adc_init(void);
-uint32_t adc_read(volatile float *val);
+uint32_t adc_read(volatile int32_t *val);
 void adc_tick(void);
 
 #endif
